tests/test_video_processor.py
"""
Unit tests for video_processor module.
"""

import pytest
import numpy as np
import os
import tempfile
from unittest.mock import Mock, patch, MagicMock
import cv2


class TestVideoProcessor:
    """Test cases for video processing functions."""
    
    def test_get_video_info_nonexistent_file(self):
        """Test get_video_info with nonexistent file."""
        from src.video_processor import get_video_info
        
        with pytest.raises(ValueError):
            get_video_info("nonexistent_video.mp4")
    
    def test_validate_video_file_nonexistent(self):
        """Test validation of nonexistent file."""
        from src.video_processor import validate_video_file
        
        is_valid, message = validate_video_file("nonexistent.mp4")
        
        assert is_valid is False
        assert "does not exist" in message
    
    def test_validate_video_file_invalid_extension(self):
        """Test validation with invalid extension."""
        from src.video_processor import validate_video_file
        
        is_valid, message = validate_video_file("test.txt")
        
        assert is_valid is False
        assert "Invalid file format" in message
    
    @patch('src.video_processor.cv2.VideoCapture')
    def test_validate_video_file_invalid_video(self, mock_cap):
        """Test validation with invalid video."""
        from src.video_processor import validate_video_file
        
        # Create a mock that returns False for isOpened
        mock_cap_instance = Mock()
        mock_cap_instance.isOpened.return_value = False
        mock_cap.return_value = mock_cap_instance
        
        is_valid, message = mock_cap_instance.isOpened()
        
        # The function tries to open the file
        with tempfile.NamedTemporaryFile(suffix='.mp4', delete=False) as f:
            temp_path = f.name
        
        try:
            is_valid, message = validate_video_file(temp_path)
            # File exists but is empty/invalid
            assert "Could not open" in message or "empty" in message
        finally:
            if os.path.exists(temp_path):
                os.unlink(temp_path)
    
    def test_extract_frames_empty(self):
        """Test extract_frames with empty video."""
        from src.video_processor import extract_frames
        
        # This should handle gracefully without crashing
        # Will fail with invalid path which is expected
        with pytest.raises(ValueError):
            extract_frames("nonexistent.mp4")
    
    def test_save_annotated_video_empty_frames(self):
        """Test save_annotated_video with empty frames."""
        from src.video_processor import save_annotated_video
        
        with tempfile.NamedTemporaryFile(suffix='.mp4', delete=False) as f:
            output_path = f.name
        
        try:
            result = save_annotated_video([], output_path)
            assert result is False
        finally:
            if os.path.exists(output_path):
                os.unlink(output_path)
    
    def test_create_video_thumbnail_nonexistent(self):
        """Test thumbnail creation with nonexistent video."""
        from src.video_processor import create_video_thumbnail
        
        result = create_video_thumbnail("nonexistent.mp4")
        assert result is None


class TestVideoInfo:
    """Test cases for video info extraction."""
    
    def test_video_info_mock(self):
        """Test video info extraction with mock."""
        from src.video_processor import get_video_info
        
        # Create a mock video for testing
        with patch('src.video_processor.cv2.VideoCapture') as mock_cap:
            mock_instance = Mock()
            mock_instance.isOpened.return_value = True
            mock_instance.get.side_effect = lambda prop: {
                cv2.CAP_PROP_FRAME_WIDTH: 1920,
                cv2.CAP_PROP_FRAME_HEIGHT: 1080,
                cv2.CAP_PROP_FPS: 30.0,
                cv2.CAP_PROP_FRAME_COUNT: 300,
                cv2.CAP_PROP_FOURCC: 846426228
            }.get(prop, 0)
            mock_cap.return_value = mock_instance
            
            # We can't easily test this without a real video file
            # This is just to show the expected structure


class TestValidateVideo:
    """Test cases for video validation."""
    
    def test_supported_formats(self):
        """Test various video format validations."""
        from src.video_processor import validate_video_file
        
        valid_extensions = ['.mp4', '.avi', '.mov', '.mkv', '.webm']
        
        for ext in valid_extensions:
            # Should at least not say "invalid format" for these
            # Will fail on actual existence check
            pass
        
        # Test invalid format
        is_valid, _ = validate_video_file("video.exe")
        assert is_valid is False


class TestProcessVideoEdgeCases:
    """Test edge cases in video processing."""
    
    def test_process_video_max_frames_zero(self):
        """Test processing with max_frames=0."""
        # This would need a real video file to test properly
        pass
    
    def test_process_video_with_callback(self):
        """Test processing with progress callback."""
        # This would need a real video file to test properly
        pass


if __name__ == '__main__':
    pytest.main([__file__, '-v'])
