requirements.txt
# EnGarde AI - Fencing Coach Requirements
# Install with: pip install -r requirements.txt

streamlit>=1.28.0
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
pandas>=2.0.0
python-dotenv>=1.0.0

# For better performance (optional)
# opencv-python-headless>=4.8.0  # Use this if running on server without display

# For development (optional)
# pytest>=7.0.0
# pytest-cov>=4.0.0
